updated dashboard changes
