+++ src/components/buyer-order.tsx
@@ -2,44 +2,43 @@ import { Button } from "./ui/button";
 
 export const BuyerOrder = () => {
   return (
-    <div className="flex items-center gap-2 container w-full p-6">
-      <table className="w-full  text-sm text-left text-gray-500 dark:text-gray-400">
-        <tbody className="">
-          <tr className="text-base text-gray-500 border-t whitespace-nowrap px-2 border-b default-border-bottom">
+    <div className="container w-full p-6 overflow-x-auto"> 
+      <table className="w-full text-sm text-left text-gray-500 dark:text-gray-400">
+        <thead>
+          <tr className="text-base text-gray-700 border-t border-b default-border-bottom">
             {orderHeader.map((header, i) => (
-              <td key={i} className="py-4 whitespace-nowrap text-center">
+              <th key={i} className="py-4 px-6 whitespace-nowrap text-center">
                 {header}
-              </td>
+              </th>
             ))}
           </tr>
-
+        </thead>
+        <tbody>
           {orders.map((order, i) => (
             <tr
               key={i}
               className="bg-white text-center border-b hover:bg-gray-50"
             >
-              <td className="text-center py-4">
-                <span className="text-lg text-gray-500 font-medium">
-                  {order.orderNumber}
-                </span>
+              <td className="py-4 px-6 text-lg text-gray-600 font-medium">
+                {order.orderNumber}
               </td>
-              <td>
+              <td className="py-4 px-6">
                 <span className="text-base text-gray-500 whitespace-nowrap">
                   {order.date}
                 </span>
               </td>
-              <td>
-                <span className="text-sm rounded text-green-500 bg-green-100 p-2">
+              <td className="py-4 px-6">
+                <span className="text-sm rounded text-green-500 bg-green-100 py-1 px-3">
                   {order.status}
                 </span>
               </td>
-              <td>
-                <span className="text-base text-gray-500 whitespace-nowrap px-2 ">
+              <td className="py-4 px-6">
+                <span className="text-base text-gray-500 whitespace-nowrap">
                   {order.amount}
                 </span>
               </td>
-              <td>
-                <Button className="text-xs shadow-none rounded-none text-black bg-main font-bold">
+              <td className="py-4 px-6">
+                <Button className="text-xs shadow-none rounded-none text-black bg-main font-bold px-4 py-2">
                   View Details
                 </Button>
               </td>
@@ -68,39 +67,10 @@ const orders: Order[] = [
     amount: "$757",
   },
   {
-    orderNumber: "#354",
-    date: "Feb 05, 2021",
-    status: "Completed",
-    amount: "$757",
+    orderNumber: "#355",
+    date: "Mar 10, 2021",
+    status: "Processing",
+    amount: "$850",
   },
-  {
-    orderNumber: "#354",
-    date: "Feb 05, 2021",
-    status: "Completed",
-    amount: "$757",
-  },
-  {
-    orderNumber: "#354",
-    date: "Feb 05, 2021",
-    status: "Completed",
-    amount: "$757",
-  },
-  {
-    orderNumber: "#354",
-    date: "Feb 05, 2021",
-    status: "Completed",
-    amount: "$757",
-  },
-  {
-    orderNumber: "#354",
-    date: "Feb 05, 2021",
-    status: "Completed",
-    amount: "$757",
-  },
-  {
-    orderNumber: "#354",
-    date: "Feb 05, 2021",
-    status: "Completed",
-    amount: "$757",
-  },
+  // ...additional order data
 ];
