+++ src/pages/user/buyer-profile/Logout.tsx
@@ -1,0 +1,7 @@
+// Logout.js
+const Logout = () => {
+    return <div>Confirm logout</div>;
+  };
+  
+  export default Logout;
+
