+++ src/pages/user/buyer-profile/Messages.tsx
@@ -1,0 +1,16 @@
+// Messages.js
+const Messages = () => (
+    <div className="p-4 w-full">
+      <h2 className="text-lg font-semibold">Messages</h2>
+      <div className="bg-white p-4 shadow rounded mt-4">
+        <div className="flex items-center justify-between">
+          <h3>Supplier X</h3>
+          <span className="text-xs text-gray-500">2 hours ago</span>
+        </div>
+        <p>Hi, we're interested in your RFQ...</p>
+      </div>
+    </div>
+  );
+  
+  export default Messages;
+
