+++ src/pages/user/buyer-profile/NotificationsIcon.tsx
@@ -1,0 +1,4 @@
+// NotificationsIcon.js
+const NotificationsIcon = () => <span>🔔</span>;
+
+export default NotificationsIcon;
