+++ src/pages/user/buyer-profile/OrderTracking.tsx
@@ -1,0 +1,11 @@
+// OrderTracking.js
+import { BuyerOrder } from "@/components/buyer-order";
+
+const OrderTracking = () => (
+  <div className="p-4">
+    <h2 className="text-lg font-semibold">Order Tracking</h2>
+    <BuyerOrder />
+  </div>
+);
+
+export default OrderTracking;
