+++ src/pages/user/buyer-profile/Ratings.tsx
@@ -1,0 +1,19 @@
+// Ratings.js
+const Ratings = () => (
+    <div className="p-4">
+      <h2 className="text-lg font-semibold">Supplier and Product Ratings</h2>
+      <div className="space-y-4 mt-4">
+        <div className="bg-white p-4 shadow rounded">
+          <h3>Supplier X - ★★★★☆</h3>
+          <p>Great communication, on-time delivery.</p>
+        </div>
+        <div className="bg-white p-4 shadow rounded">
+          <h3>Supplier Y - ★★★☆☆</h3>
+          <p>Product quality was okay, but shipping was delayed.</p>
+        </div>
+      </div>
+    </div>
+  );
+  
+  export default Ratings;
+
